% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2bibtex.R
\name{csv2bibtex}
\alias{csv2bibtex}
\title{csv2bibtex Function to convert a csv dataframe to BibTeX format}
\usage{
csv2bibtex(csv_df, save_name)
}
\arguments{
\item{csv_df}{CSV dataframe to convert to bibTex}

\item{save_name}{user input name to save file}
}
\value{
saves a bibTex file in working directory
}
\description{
csv2bibtex Function to convert a csv dataframe to BibTeX format
}
\author{
Emmanuel Adeyemo, Kelechi Igwe
}
