% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_articles.R
\name{plot_core_articles}
\alias{plot_core_articles}
\title{plot_core_articles line plot of core articles}
\usage{
plot_core_articles(df_journal)
}
\arguments{
\item{df_journal}{dataframe with similar or conflicted picks from both reviewers. Output from both_selected or get_conflict function.}
}
\value{
ggplot image
}
\description{
plot_core_articles line plot of core articles
}
\author{
Emmanuel Adeyemo, Kelechi Igwe
}
