% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_core_articles.R
\name{plot_core_by_year}
\alias{plot_core_by_year}
\title{plot_core_by_year line plot of core articles by year}
\usage{
plot_core_by_year(df_journal)
}
\arguments{
\item{df_journal}{dataframe with similar or conflicted picks from both reviewers. Output from both_selected or get_conflict function.}
}
\value{
ggplot image
}
\description{
plot_core_by_year line plot of core articles by year
}
\author{
Emmanuel Adeyemo, Kelechi Igwe
}
